refactor(sqlite): extract shared dtype/type-map into _sqlite_typemap
